// frontend/src/shared/api/generated/models/eventSignupStatsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CodigoActivo.API
 * OpenAPI spec version: 1.0
 */
import type { EventSignupStatsActivityResponse } from './eventSignupStatsActivityResponse';
import type { EventSignupStatsRoleResponse } from './eventSignupStatsRoleResponse';
import type { EventSignupStatsStatusResponse } from './eventSignupStatsStatusResponse';
import type { EventSignupStatsTotalsResponse } from './eventSignupStatsTotalsResponse';

export interface EventSignupStatsResponse {
  eventId?: string;
  /** @nullable */
  roles?: EventSignupStatsRoleResponse[] | null;
  /** @nullable */
  statuses?: EventSignupStatsStatusResponse[] | null;
  /** @nullable */
  activities?: EventSignupStatsActivityResponse[] | null;
  totals?: EventSignupStatsTotalsResponse;
}